+++ NEWS.md
@@ -3,7 +3,7 @@
 - add loadFunctions()
 - multiColumnLookup()
   - allow to lookup more than one value column
-  - add argument "drop" 
+  - add arguments "drop", "includeKeys" 
 
 # [kwb.utils 0.14.1](https://github.com/KWB-R/kwb.utils/releases/tag/v0.14.0) <small>2023-04-21</small>